Fixed x64 compilation with MSVS

+++ include/cryptoplus/asn1/string.hpp
@@ -245,7 +245,7 @@ namespace cryptoplus
 		}
 		inline void string::set_data(const void* _data, size_t data_len)
 		{
-			error::throw_error_if_not(ASN1_STRING_set(ptr().get(), _data, data_len) != 0);
+			error::throw_error_if_not(ASN1_STRING_set(ptr().get(), _data, static_cast<int>(data_len)) != 0);
 		}
 		inline void string::set_data(const char* _data)
 		{
